+++ src/obf_run.h
@@ -8,9 +8,9 @@ obf_run_obfuscate(const mmap_vtable *mmap, const obfuscator_vtable *vt,
                   size_t *kappa, size_t nthreads, aes_randstate_t rng);
 
 int
-obf_run_evaluate(const mmap_vtable *mmap, const obfuscator_vtable *vt, 
-                 const char *fname, obf_params_t *op, const int *input,
-                 size_t ninputs, int *output, size_t noutputs, size_t nthreads,
+obf_run_evaluate(const mmap_vtable *mmap, const obfuscator_vtable *vt,
+                 const char *fname, obf_params_t *op, const long *input,
+                 size_t ninputs, long *output, size_t noutputs, size_t nthreads,
                  size_t *kappa, size_t *npowers);
 
 size_t
